pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "eventlet>=0.40.1",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "netifaces>=0.10.6",
    "paramiko>=3.5.1",
    "psycopg2-binary>=2.9.10",
    "python-nmap>=0.7.1",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "psutil>=7.0.0",
    "flask-socketio>=5.5.1",
]
